wgengine/router/dns: move to net/dns.
Preparation for merging the APIs and whatnot.
