+++ net/dns/manager_linux.go
