+++ net/dns/manager_openbsd.go
