+++ net/dns/manager_windows.go
