+++ net/dns/noop.go
