+++ net/dns/registry_windows.go
